hw 6.3 done 6.4 v1

+++ 06_InheritanceAndPolymorphism/BankClient/src/NegativeRemnantException.java
@@ -2,7 +2,7 @@ import java.math.BigDecimal;
 
 public class NegativeRemnantException extends Exception {
 
-    public NegativeRemnantException(double balance, double sum) {
+    public NegativeRemnantException(BigDecimal balance, BigDecimal sum) {
         super("Недостаточно средств на счёте! \n Баланс: " + balance + "\n Сумма: " + sum);
     }
 }
